FileAnalysis: update baselines.
(Yes, I spent a day checking over the diffs.)

The only interesting thing to note that I can remember is that there
was the old baseline may have had MD5's calculated for HTTP
partial content, but it's not anymore becase the FAF isn't yet
attempting to do any reassembly of non-linear file transfers.
(And the old way of always generating hashes for partial content
would be wrong in a lot of cases).